Refactoring for manager view
